feat: set and notice for account expiration #397
