+++ src/api/bkuser_core/config/common/system.py
@@ -82,6 +82,9 @@ DRF_CROWN_DEFAULT_CONFIG = {"remain_request": True}
 # 复用 API, 接口参数中存在 SYNC_API_PARAM 时, 以sync的接口协议返回
 SYNC_API_PARAM = "for_sync"
 
+# 通知发送时间间隔
+NOTICE_INTERVAL = 15
+
 
 # ==============================================================================
 # 开发调试
