+++ src/api/bkuser_core/profiles/account_expiration_notifier.py
@@ -1,0 +1,175 @@
+# -*- coding: utf-8 -*-
+"""
+TencentBlueKing is pleased to support the open source community by making 蓝鲸智云-用户管理(Bk-User) available.
+Copyright (C) 2017-2021 THL A29 Limited, a Tencent company. All rights reserved.
+Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
+You may obtain a copy of the License at http://opensource.org/licenses/MIT
+Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+specific language governing permissions and limitations under the License.
+"""
+import datetime
+import logging
+
+from django.db.models import Exists, OuterRef, Q
+
+from bkuser_core.audit.models import LogIn
+from bkuser_core.categories.constants import CategoryType
+from bkuser_core.categories.models import ProfileCategory
+from bkuser_core.common.notifier import send_mail, send_sms
+from bkuser_core.user_settings.constants import ACCOUNT_EXPIRATION_NOTICE_INTERVAL_META_KEY
+from bkuser_core.profiles.models import Profile
+from bkuser_core.user_settings.constants import SettingsEnableNamespaces
+from bkuser_core.user_settings.loader import ConfigProvider
+from bkuser_core.user_settings.models import Setting
+
+logger = logging.getLogger(__name__)
+
+EMAIL_NOTICE_METHOD = "send_email"
+SMS_NOTICE_METHOD = "send_sms"
+
+
+def get_profiles_for_account_expiration():
+    """
+    获取 需要进行账号过期相关通知的用户
+    """
+
+    profiles = []
+    category_ids = ProfileCategory.objects.filter(type=CategoryType.LOCAL.value).values_list("id", flat=True)
+
+    for category_id in category_ids:
+        notice_interval = Setting.objects.filter(
+            category_id=category_id,
+            meta__key=ACCOUNT_EXPIRATION_NOTICE_INTERVAL_META_KEY,
+            meta__namespace=SettingsEnableNamespaces.ACCOUNT.value
+        ).first().value
+
+        expiration_times = get_expiration_dates(notice_interval)
+        logined_profiles = get_logined_profiles()
+
+        expiring_profiles = logined_profiles.filter(
+            Q(account_expiration_date__lt=datetime.date.today()) | Q(account_expiration_date__in=expiration_times),
+            category_id=category_id)
+        profiles.extend(expiring_profiles)
+    return profiles
+
+
+def get_expiration_dates(notice_interval):
+    """
+    获取需要进行通知的 过期时间列表
+    """
+    expiration_times = []
+    for t in notice_interval:
+        expiration_time = datetime.date.today() + datetime.timedelta(days=t)
+        expiration_times.append(expiration_time)
+
+    return expiration_times
+
+
+def get_logined_profiles():
+    """
+    获取在平台登录过的所有用户
+    """
+    subquery = LogIn.objects.filter(profile=OuterRef('pk')).values_list('id')
+    logined_profile_ids = Profile.objects.annotate(
+        temp=Exists(subquery)).filter(temp=True).values_list('id', flat=True)
+    logined_profiles = Profile.objects.filter(id__in=logined_profile_ids)
+
+    return logined_profiles
+
+
+def get_notice_config_for_account_expiration(profile):
+    """
+    整合 账号过期 通知内容
+    """
+    notice_config = {}
+    expire_at = datetime.date.today() - profile.account_expiration_date
+
+    config_loader = ConfigProvider(profile.category_id)
+    notice_methods = config_loader["account_expiration_notice_methods"]
+
+    if not notice_methods:
+        return
+
+    if EMAIL_NOTICE_METHOD in notice_methods:
+        email_config = (
+            config_loader["expired_account_email_config"]
+            if expire_at.days < 0
+            else config_loader["expiring_account_email_config"]
+        )
+
+        message = (
+            email_config["content"].format(username=profile.username)
+            if expire_at.days < 0
+            else email_config["content"].format(
+                username=profile.username, expire_at=expire_at.days
+            )
+        )
+
+        notice_config.update(
+            {
+                "send_email":
+                    {
+                        "sender": email_config["sender"],
+                        "receiver": [profile.email],
+                        "message": message,
+                        "title": email_config["title"]
+                    }
+
+            }
+        )
+
+    if SMS_NOTICE_METHOD in notice_methods:
+        sms_config = (
+            config_loader["expired_account_sms_config"]
+            if expire_at.days < 0
+            else config_loader["expiring_account_sms_config"]
+        )
+
+        message = (
+            sms_config["content"].format(username=profile.username)
+            if expire_at.days < 0
+            else sms_config["content"].format(
+                username=profile.username, expire_at=expire_at.days
+            )
+        )
+
+        notice_config.update(
+            {
+                "send_sms":
+                    {
+                        "receivers": [profile.telephone],
+                        "message": message
+                    }
+
+            }
+        )
+
+    return notice_config
+
+
+class AccountExpirationNotifier:
+    def handler(self, notice_config):
+
+        notice_method_map = {
+            "send_email": self._notice_by_email,
+            "send_sms": self._notice_by_sms,
+        }
+
+        for notice_method in notice_config:
+            notice_method_map[notice_method](notice_config[notice_method])
+
+    def _notice_by_email(self, email_config):
+        send_mail(
+            sender=email_config["sender"],
+            receivers=email_config["receivers"],
+            message=email_config["message"],
+            title=email_config["title"],
+        )
+
+    def _notice_by_sms(self, sms_config):
+        send_sms(
+            sender=sms_config["sender"],
+            receivers=sms_config["receivers"],
+            message=sms_config["message"]
+        )
