+++ src/api/bkuser_core/profiles/models.py
@@ -104,7 +104,8 @@ class Profile(TimestampedModel):
     # ----------------------- 职位相关 -----------------------
 
     # ----------------------- 账号相关 -----------------------
-    account_expiration_time = models.DateField(verbose_name=_("账号过期时间"), null=True, blank=True)
+    account_expiration_date = models.DateField(verbose_name=_("账号过期时间"), null=True, blank=True,
+                                               default=datetime.date(year=2100, month=1, day=1))
     # ----------------------- 账号相关 -----------------------
 
     # ----------------------- 国际化相关 -----------------------
@@ -151,7 +152,7 @@ class Profile(TimestampedModel):
         verbose_name = "用户信息"
         verbose_name_plural = "用户信息"
         unique_together = ("username", "category_id")
-        index_together = ("category_id", "account_expiration_time")
+        index_together = ("category_id", "account_expiration_date")
 
     def custom_validate(self):
         validate_domain(self.domain)
